Refactoring GNB.tsx
